+++ src/components/Profile/ProfileSidebar.tsx
@@ -1,6 +1,8 @@
 import React from "react";
 import Image from "next/image";
-
+import user from "../../../public/user.png";
+import car from "../../../public/car.png";
+import checkbox from "../../../public/checkbox.png";
 type ProfileSidebarProps = {
   option: "user" | "carpool" | "account";
   setOption: React.Dispatch<
@@ -10,29 +12,25 @@ type ProfileSidebarProps = {
 
 const ProfileSidebar = ({ option, setOption }: ProfileSidebarProps) => {
   const baseButton =
-    "px-4 py-2 relative items-center  flex gap-2 text-northeastern-red font-montserrat text-xl  ";
+    "px-3 py-2 relative items-center  flex gap-2 text-northeastern-red font-montserrat text-xl lg:text-2xl  ";
   const selectedButton = " font-bold ";
   return (
-    <div className="h-full w-full  ">
-      <div className="mt-6 flex w-full flex-col items-start  justify-center gap-4 lg:ml-12  lg:text-start">
+    <div className="my-16 h-full w-full  ">
+      <div className="mt-6 flex w-full flex-col items-start  justify-center gap-6 lg:ml-10  lg:text-start">
         <button
           className={baseButton + (option === "user" && selectedButton)}
           onClick={() => setOption("user")}
         >
           <div
-            className={`relative  ${
+            className={`relative  w-12  ${
               option === "user" &&
               " after:absolute after:inset-0 after:bg-northeastern-red after:mix-blend-screen "
             }
              `}
-            style={{ width: "31.95px", height: "34.65px" }}
           >
-            <Image
-              src="/user.png"
-              alt="user"
-              className="absolute inset-0 h-full w-full object-cover grayscale"
-              layout={"fill"}
-            />
+            <div className="relative flex justify-center  ">
+              <Image src={user} alt="user" />
+            </div>
           </div>
           User Profile
         </button>
@@ -41,19 +39,15 @@ const ProfileSidebar = ({ option, setOption }: ProfileSidebarProps) => {
           onClick={() => setOption("carpool")}
         >
           <div
-            className={`relative  ${
+            className={`relative w-12  ${
               option === "carpool" &&
               " after:absolute after:inset-0 after:bg-northeastern-red after:mix-blend-screen "
             }
              `}
-            style={{ width: "38.25px", height: "32.2px" }}
           >
-            <Image
-              src="/car.png"
-              alt="car"
-              className="absolute inset-0 h-full w-full object-cover grayscale"
-              layout={"fill"}
-            />
+            <div className="relative flex justify-center  ">
+              <Image src={car} alt="car" />
+            </div>
           </div>
           Carpool Details
         </button>
@@ -62,19 +56,15 @@ const ProfileSidebar = ({ option, setOption }: ProfileSidebarProps) => {
           onClick={() => setOption("account")}
         >
           <div
-            className={`relative  ${
+            className={`relative w-12  ${
               option === "account" &&
               " after:absolute after:inset-0 after:bg-northeastern-red after:mix-blend-screen "
             }
              `}
-            style={{ width: "34.5px", height: "34.5px" }}
           >
-            <Image
-              src="/checkbox.png"
-              alt="checkbox"
-              className="absolute inset-0 h-full w-full object-cover grayscale"
-              layout={"fill"}
-            />
+            <div className="relative flex justify-center  ">
+              <Image src={checkbox} alt="checkbox" />
+            </div>
           </div>
           Account Status
         </button>
