--- src/pages/profile/profile2.tsx
@@ -1,257 +1,0 @@
-import React, { useCallback, useEffect, useState } from "react";
-import { FieldErrors, useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { GetServerSidePropsContext, NextPage } from "next";
-import { useRouter } from "next/router";
-import { toast } from "react-toastify";
-import { getSession, useSession } from "next-auth/react";
-import { trpc } from "../../utils/trpc";
-import { CarpoolAddress, OnboardingFormInputs } from "../../utils/types";
-import {
-  onboardSchema,
-  profileDefaultValues,
-} from "../../utils/profile/zodSchema";
-
-import Spinner from "../../components/Spinner";
-import InitialStep from "../../components/Setup/InitialStep";
-import { FaArrowRight } from "react-icons/fa";
-import StepTwo from "../../components/Setup/StepTwo";
-import ProgressBar from "../../components/Setup/ProgressBar";
-import StepThree from "../../components/Setup/StepThree";
-import { SetupContainer } from "../../components/Setup/SetupContainer";
-import StepFour from "../../components/Setup/StepFour";
-import { Role } from "@prisma/client";
-import {
-  trackFTUECompletion,
-  trackFTUEStep,
-  trackProfileCompletion,
-} from "../../utils/mixpanel";
-import { useUploadFile } from "../../utils/profile/useUploadFile";
-import { useAddressSelection } from "../../utils/useAddressSelection";
-import {
-  updateUser,
-  useEditUserMutation,
-} from "../../utils/profile/updateUser";
-
-import ProfileSidebar from "../../components/Profile/ProfileSidebar";
-import UserSection from "../../components/Profile/UserSection";
-import Header from "../../components/Header";
-import CarpoolSection from "../../components/Profile/CarpoolSection";
-import AccountSection from "../../components/Profile/AccountSection";
-
-export async function getServerSideProps(context: GetServerSidePropsContext) {
-  const session = await getSession(context);
-  if (!session?.user) {
-    return {
-      redirect: {
-        destination: "/sign-in",
-        permanent: false,
-      },
-    };
-  }
-
-  return {
-    props: {},
-  };
-}
-const Profile2: NextPage = () => {
-  const router = useRouter();
-  const [option, setOption] = useState<"user" | "carpool" | "account">("user");
-  const [isLoading, setIsLoading] = useState(false);
-  const [initialLoad, setInitialLoad] = useState(true);
-  const [selectedFile, setSelectedFile] = useState<File | null>(null);
-  const { uploadFile } = useUploadFile(selectedFile);
-  const { data: session } = useSession();
-  const { data: user } = trpc.user.me.useQuery(undefined, {
-    refetchOnMount: true,
-  });
-  const editUserMutation = useEditUserMutation(
-    router,
-    () => setIsLoading(false),
-    false
-  );
-  const startAddressHook = useAddressSelection();
-  const companyAddressHook = useAddressSelection();
-
-  const { setSelectedAddress: setStartAddressSelected } = startAddressHook;
-  const { setSelectedAddress: setCompanyAddressSelected } = companyAddressHook;
-
-  useEffect(() => {
-    if (user?.startAddress && user.startAddress !== "") {
-      setStartAddressSelected({
-        place_name: user.startAddress,
-        center: [user.startCoordLng, user.startCoordLat],
-      });
-    }
-    if (user?.companyAddress && user.companyAddress !== "") {
-      setCompanyAddressSelected({
-        place_name: user.companyAddress,
-        center: [user.companyCoordLng, user.companyCoordLat],
-      });
-    }
-  }, [user, setStartAddressSelected, setCompanyAddressSelected]);
-
-  const {
-    register,
-    setValue,
-    formState: { errors },
-    watch,
-    handleSubmit,
-    reset,
-    control,
-  } = useForm<OnboardingFormInputs>({
-    mode: "onChange",
-    defaultValues: profileDefaultValues,
-    resolver: zodResolver(onboardSchema),
-  });
-
-  useEffect(() => {
-    if (initialLoad && user) {
-      reset({
-        role: user.role,
-        seatAvail: user.seatAvail,
-        status: user.status,
-        companyName: user.companyName,
-        companyAddress: user.companyAddress,
-        startAddress: user.startAddress,
-        preferredName: user.preferredName,
-        pronouns: user.pronouns,
-        daysWorking: user.daysWorking
-          ? user.daysWorking.split(",").map((bit) => bit === "1")
-          : profileDefaultValues.daysWorking,
-        startTime: user.startTime!,
-        endTime: user.endTime!,
-        coopStartDate: user.coopStartDate!,
-        coopEndDate: user.coopEndDate!,
-        bio: user.bio,
-      });
-      setInitialLoad(false);
-    }
-  }, [initialLoad, reset, user]);
-  const role = watch("role");
-
-  useEffect(() => {
-    const seatAvail = watch("seatAvail");
-    if (role === Role.DRIVER && seatAvail <= 0) {
-      setValue("seatAvail", 1);
-    } else if (role !== Role.DRIVER) {
-      setValue("seatAvail", 0);
-    }
-  }, [setValue, watch, role]);
-
-  const onSubmit = async (values: OnboardingFormInputs) => {
-    setIsLoading(true);
-    const userInfo = {
-      ...values,
-      companyCoordLng: companyAddressHook.selectedAddress.center[0],
-      companyCoordLat: companyAddressHook.selectedAddress.center[1],
-      startCoordLng: startAddressHook.selectedAddress.center[0],
-      startCoordLat: startAddressHook.selectedAddress.center[1],
-      seatAvail: values.role === "RIDER" ? 0 : values.seatAvail,
-    };
-    if (selectedFile) {
-      try {
-        await uploadFile();
-      } catch (error) {
-        console.error("File upload failed:", error);
-      }
-    }
-    const sessionName = session?.user?.name ?? "";
-    try {
-      await updateUser({
-        userInfo,
-        sessionName,
-        mutation: editUserMutation,
-      });
-      trackProfileCompletion(userInfo.role, userInfo.status);
-      toast.success("User profile updated successfully!");
-    } catch (error) {
-      toast.error("Failed to update user profile. Please try again.");
-    } finally {
-      setIsLoading(false);
-    }
-  };
-  const onError = (errors: FieldErrors<OnboardingFormInputs>) => {
-    const firstErrorKey = Object.keys(errors)[0];
-
-    if (firstErrorKey) {
-      if (
-        ["preferredName", "pronouns", "role", "bio", "seatAvail"].includes(
-          firstErrorKey
-        )
-      ) {
-        setOption("user");
-      } else if (
-        [
-          "startAddress",
-          "companyAddress",
-          "startTime",
-          "endTime",
-          "daysWorking",
-        ].includes(firstErrorKey)
-      ) {
-        setOption("carpool");
-      } else {
-        setOption("account");
-      }
-    }
-    toast.error("One or more fields are invalid, please fix and try again.");
-  };
-  if (isLoading || !user) {
-    return (
-      <div className="fixed inset-0 z-50 flex items-center justify-center bg-white ">
-        <Spinner />
-      </div>
-    );
-  }
-  return (
-    <div className="relative h-full w-full overflow-hidden">
-      <Header />
-      <div className="grid h-[91.5%] grid-cols-[250px_repeat(2,1fr)] lg:grid-cols-[350px_repeat(2,1fr)]">
-        {/* Sidebar */}
-        <div className="col-start-1 col-end-2 w-[250px] border-r-4 border-busy-red bg-stone-100 lg:w-[350px]">
-          <ProfileSidebar option={option} setOption={setOption} />
-        </div>
-
-        {/* Main Section */}
-        <div className="col-start-2 col-end-4 flex shrink items-start justify-center overflow-hidden">
-          <div className="overflow-hidden  lg:-translate-x-[175px]  lg:overflow-visible">
-            {option === "user" ? (
-              <UserSection
-                watch={watch}
-                onFileSelect={setSelectedFile}
-                errors={errors}
-                register={register}
-                onSubmit={handleSubmit(onSubmit, onError)}
-                setValue={setValue}
-              />
-            ) : option === "carpool" ? (
-              <CarpoolSection
-                watch={watch}
-                onFileSelect={setSelectedFile}
-                errors={errors}
-                register={register}
-                setValue={setValue}
-                onSubmit={handleSubmit(onSubmit, onError)}
-                startAddressHook={startAddressHook}
-                companyAddressHook={companyAddressHook}
-                control={control}
-              />
-            ) : option === "account" ? (
-              <AccountSection
-                control={control}
-                watch={watch}
-                onSubmit={handleSubmit(onSubmit, onError)}
-                errors={errors}
-                setValue={setValue}
-              />
-            ) : (
-              <></>
-            )}
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
-export default Profile2;
